requirements.txt
Django>=3.0,<4.0
djangorestframework
psycopg2>=2.8
social-auth-app-django
django-environ
django-cors-headers
authlib ~= 1.0
python-dotenv ~= 0.19
requests ~= 2.27
faker